Migrate xbpm-randmat counts/indices to size_t; update matrix APIs and fix printf formats
- Convert site/count/index types to size_t to match allocation APIs and avoid signed/unsigned issues.
  - Add #include <stddef.h> and change nsites / idx / ord_sites types in prm_def.h.
  - Update functions, loops and prototypes to use size_t:
    - data_read.c
    - index helpers (index_order_by_position, exchange_values, site_visited, min_and_max) now use size_t.
    - roi_indexation uses size_t for allocations/counters and copies.
    - data_read loop counter uses size_t.
    - Update debug/commented prints to use %zu.
    - matrix_read loop uses size_t and prints line numbers with %zu.

  - parameters_read.c
    - parse -n into prm.nsites using strtoul() (size_t).

  - positions_calc.c
    - loops and index uses converted to size_t; cast roi->nsites to
      double where used in arithmetic (avoid integer promotion).
    - update commented debug prints to %zu.

  - positions_print.c
    - loop and index types updated to size_t.

  - random_walk.c
    - ROI-aware functions use size_t (chi2_calc now guards roi->nsites <= 1 and casts for division).
    - Update commented debug prints to use %zu.

- Matrix operations & prototypes hardened:
  - matrix_operations.h
    - change matrix/vector dimension parameters to size_t (matrix_transpose, matrix_product, matrix_vector_product, dot_product, vector_sum).
  - matrix_operations.c
    - update implementations to use size_t for dimensions/indices and corrected row-major flat-array indexing.
    - fixed stray brace and cleaned ROI helpers to use size_t indices.

- Main / test cleanup:
  - main.c
    - use size_t for local loop counters and matrix_show signature.
    - updated calls to new data_read(prm) prototype.

- main_test_arrays.c deleted.

- Housekeeping & safety:
  - Replaced many %d debug/print usages that print indices with %zu.
  - Added guards to avoid unsigned underflow/division by zero (chi2_calc).
  - Kept changes focused on the xbpm-randmat area (src is a link; no duplicate edits required).

Notes / Impact
  - These edits change several public/internal function signatures (matrix ops prototypes, data_read), so callers must be compiled/updated accordingly. This is a breaking change for C callers that still use int-typed signatures.
  - I did not run a full build in this commit; a compile pass is the next step to catch remaining warnings and call-site mismatches.
  - Remaining work: complete a full sweep for printf format specifiers, update any remaining call sites across the repo, add overflow checks on multiplications used for allocations, and run/verify a full compilation and tests.

- Files changed (high level)

 new file:   Makefile
 modified:   data_read.c
 modified:   main.c
 deleted:    main_test_arrays.c
 modified:   matrix_operations.c
 modified:   matrix_operations.h
 modified:   parameters_read.c
 modified:   positions_calc.c
 modified:   positions_print.c
 modified:   prm_def.h
 modified:   random_walk.c

+++ positions_print.c
@@ -7,8 +7,8 @@
  */
 void positions_print(dataset * ds, double * hh, double * vv)
 {
-    int ia;
-    for (int ii = 0; ii < ds->nsites; ii++)
+    size_t ia;
+    for (size_t ii = 0; ii < ds->nsites; ii++)
     {
         ia = ds->ord_sites[ii];
         printf("%12.4f %12.4f %12.4f %12.4f \n",
